Refactor plots to their own function
