+++ src/gemdat/plots/matplotlib/__init__.py
@@ -2,34 +2,29 @@
 matplotlib."""
 from __future__ import annotations
 
-from ._displacements import (
-    displacement_histogram,
-    displacement_per_atom,
-    displacement_per_element,
-    msd_per_element,
-)
-from ._jumps import (
-    collective_jumps,
-    jumps_3d,
-    jumps_3d_animation,
-    jumps_vs_distance,
-    jumps_vs_time,
-)
-from ._paths import (
-    energy_along_path,
-    path_on_grid,
-)
-from ._rdf import radial_distribution
-from ._orientations import (
-    autocorrelation,
-    bond_length_distribution,
-    rectilinear,
-)
+from ._displacement_histogram import displacement_histogram
+from ._displacement_per_atom import displacement_per_atom
+from ._displacement_per_element import displacement_per_element
+from ._msd_per_element import msd_per_element
+
+from ._collective_jumps import collective_jumps
+from ._jumps_3d import jumps_3d
+from ._jumps_3d_animation import jumps_3d_animation
+from ._jumps_vs_distance import jumps_vs_distance
+from ._jumps_vs_time import jumps_vs_time
+
+from ._energy_along_path import energy_along_path
+from ._path_on_grid import path_on_grid
+from ._radial_distribution import radial_distribution
+
+from ._autocorrelation import autocorrelation
+from ._bond_length_distribution import bond_length_distribution
+from ._rectilinear import rectilinear
+
 from ._shape import shape
-from ._vibration import (
-    frequency_vs_occurence,
-    vibrational_amplitudes,
-)
+
+from ._frequency_vs_occurence import frequency_vs_occurence
+from ._vibrational_amplitudes import vibrational_amplitudes
 
 __all__ = [
     'bond_length_distribution',
