+++ src/gemdat/plots/matplotlib/_displacement_per_atom.py
@@ -1,0 +1,30 @@
+from __future__ import annotations
+
+import matplotlib.pyplot as plt
+
+from gemdat.trajectory import Trajectory
+
+
+def displacement_per_atom(*, trajectory: Trajectory) -> plt.Figure:
+    """Plot displacement per atom.
+
+    Parameters
+    ----------
+    trajectory : Trajectory
+        Input trajectory, i.e. for the diffusing atom
+
+    Returns
+    -------
+    fig : matplotlib.figure.Figure
+        Output figure
+    """
+    fig, ax = plt.subplots()
+
+    for distances in trajectory.distances_from_base_position():
+        ax.plot(distances, lw=0.3)
+
+    ax.set(title='Displacement per site',
+           xlabel='Time step',
+           ylabel='Displacement (Å)')
+
+    return fig
